+++ src/configs/tailwind/index.ts
@@ -1,5 +1,5 @@
-import { generateModuleHTMLComponent, generateModuleHTMLSnippet } from "../generateModuleComponents"
-import type { ModuleConfig } from "../moduleConfigs"
+import { generateModuleHTMLComponent, generateModuleHTMLSnippet } from "../../generators/generateModuleComponents"
+import type { PackageConfig } from "../index"
 
 const tailwindDemoComponent = `<template>
   ${generateModuleHTMLComponent(
@@ -14,7 +14,7 @@ const tailwindDemoComponent = `<template>
 </template>
 `
 
-const tailwind: ModuleConfig = {
+const tailwind: PackageConfig = {
   humanReadableName: "Tailwind CSS",
   description: "A utility-first CSS framework packed with classes that can be composed to build any design, directly in your markup. See more: https://tailwindcss.com/",
   dependencies: [{
