+++ src/generators/generateModuleComponents.ts
